clang-tidy --fix

+++ source/Common/include/TMIV/Common/Bitstream.h
@@ -101,7 +101,7 @@ public:
 
   template <typename Integer> void writeBits(const Integer &value, unsigned bits);
 
-  void putFlag(bool value) { writeBits(int{value}, 1); }
+  void putFlag(bool value) { writeBits(int{static_cast<int>(value)}, 1); }
   void putUint8(uint8_t value) { writeBits(value, 8); }
   void putUint16(uint16_t value) { writeBits(value, 16); }
   void putUint32(uint32_t value) { writeBits(value, 32); }
